refactor: station 패키지 리펙터링
