+++ src/main/java/nextstep/subway/station/StationResponse.java
@@ -1,22 +1,14 @@
 package nextstep.subway.station;
 
+import lombok.Getter;
+
+@Getter
 public class StationResponse {
-    private Long id;
-    private String name;
+	private Long id;
+	private String name;
 
-    public StationResponse() {
-    }
-
-    public StationResponse(Long id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
+	public StationResponse(Station station) {
+		this.id = station.getId();
+		this.name = station.getName();
+	}
 }
